requirements.txt
uvicorn==0.23.2
argon2-cffi==23.1.0
fastapi==0.104.0
graphene==3.3
psycopg2 == 2.9.9
SQLAlchemy==2.0.22
starlette==0.27.0
starlette-graphene3==0.6.0
requests==2.31.0
pyotp==2.9.0
PyJWT==2.8.0
python-dotenv==1.0.0
